Moving DTOs into folders
