+++ src/main/java/com/danielagapov/spawn/DTOs/ChatMessage/ChatMessageLikesDTO.java
@@ -1,4 +1,4 @@
-package com.danielagapov.spawn.DTOs;
+package com.danielagapov.spawn.DTOs.ChatMessage;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
