+++ src/main/java/com/danielagapov/spawn/Services/Auth/AuthService.java
@@ -1,6 +1,6 @@
 package com.danielagapov.spawn.Services.Auth;
 
-import com.danielagapov.spawn.DTOs.AuthUserDTO;
+import com.danielagapov.spawn.DTOs.User.AuthUserDTO;
 import com.danielagapov.spawn.DTOs.User.FullUserDTO;
 import com.danielagapov.spawn.DTOs.User.UserDTO;
 import com.danielagapov.spawn.Exceptions.EmailAlreadyExistsException;
